Moving keyword logic into separate helper file
