+++ medicines/web/src/services/azure-search.ts
@@ -1,12 +1,11 @@
 import fetch, { Response } from 'node-fetch';
+import { buildFuzzyQuery } from './keyword';
 
 const searchApiVersion = process.env.AZURE_SEARCH_API_VERSION;
-const searchExactnessBoost = process.env.AZURE_SEARCH_EXACTNESS_BOOST;
 const searchIndex = process.env.AZURE_SEARCH_INDEX;
 const searchKey = process.env.AZURE_SEARCH_KEY;
 const searchScoringProfile = process.env.AZURE_SEARCH_SCORING_PROFILE;
 const searchService = process.env.AZURE_SEARCH_SERVICE;
-const searchWordFuzziness = process.env.AZURE_SEARCH_WORD_FUZZINESS;
 
 export enum DocType {
   Par = 'Par',
@@ -36,43 +35,6 @@ export interface ISearchResults {
   resultCount: number;
   results: ISearchResult[];
 }
-
-const extractProductLicenseRegExp: RegExp = new RegExp(
-  '(\\b|PL)(\\s+|/|_|-)*(\\d{5})(\\s+|/|_|-)*(\\d{4})',
-  'ig',
-);
-
-const escapeSpecialWords = (word: string): string =>
-  word.replace(/(\|\||&&|\bAND\b|\bOR\b|\bNOT\b)/gi, `\\$1`);
-
-const preferExactMatchButSupportFuzzyMatch = (word: string): string =>
-  `${word}~${searchWordFuzziness} ${word}^${searchExactnessBoost}`;
-
-const addNormalizedProductLicenses = (q: string): string => {
-  const normalizedProductLicences = q
-    .match(extractProductLicenseRegExp)
-    ?.map(match => match.replace(extractProductLicenseRegExp, 'PL$3$5'));
-
-  if (normalizedProductLicences && normalizedProductLicences.length) {
-    const normalizedProductLicencesString: string = normalizedProductLicences.join(
-      ' ',
-    );
-    return `${q} ${normalizedProductLicencesString}`;
-  }
-
-  return `${q}`;
-};
-
-const splitByNonSearchableCharacters = (query: string) =>
-  query.split(/(?:[,+\-!(){}\[\]^~*?:\/]|\s+)/gi);
-
-const buildFuzzyQuery = (query: string): string => {
-  return splitByNonSearchableCharacters(addNormalizedProductLicenses(query))
-    .filter(x => x.length > 0)
-    .map(word => escapeSpecialWords(word))
-    .map(word => preferExactMatchButSupportFuzzyMatch(word))
-    .join(' ');
-};
 
 const calculatePageStartRecord = (page: number, pageSize: number): number =>
   pageSize * (page - 1);
